refactor store - break out into substores
