+++ src/components/BoardTile.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="group relative flex-none" :style="tile.styles"
-    :class="tile.classes, store.selectedPawn?.tile.id == tile.id && 'bg-opacity-90'" :key="tile.id">
+    :class="tile.classes, boardStore.selectedPawn?.tile.id == tile.id && 'bg-opacity-90'" :key="tile.id">
     <div v-if="!tile.isCornerTile" class="flex flex-col h-full w-full">
       <FighterPawn v-if="fighterOnThisTile" :fighter="fighterOnThisTile" :tile="tile" class="relative z-20" />
     </div>
@@ -11,17 +11,17 @@
 import FighterPawn from "../components/FighterPawn.vue";
 import Tile from '../models/Tile';
 import { storeToRefs } from 'pinia';
-import { useStore } from '../store';
+import { useBoardStore } from '../store';
 import { computed } from "vue";
 
 const props = defineProps<{ tile: Tile }>()
 
-const store = useStore()
+const boardStore = useBoardStore()
 
-const { fightersOnTiles } = storeToRefs(store)
+const { fightersKeyedByTileId } = storeToRefs(boardStore)
 
 const fighterOnThisTile = computed(() => {
-  const fighter = fightersOnTiles.value[props.tile.id]
+  const fighter = fightersKeyedByTileId.value[props.tile.id]
 
   return fighter?.isAlive ? fighter : null
 })
